Create load-imgs
This allows the image files to be downloaded and decrypted and placed in an appropriate folder. 
To follow the process more closely, the number of files found is determined at the beginning and the current status is output during processing. 

For the process to work, the corresponding URL, Access Key, Secret Key, Bucket Name and Encryption Key still need to be inserted. The destination folder of the files should also be changed if necessary.